Move Footer and Header styles to another file.
