+++ src/main/layout/footer/Footer.styles.js
@@ -1,0 +1,16 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles({
+    root: {
+        padding: "30px",
+        marginTop: 'auto',
+    },
+    footer_title: {
+        fontWeight: "bold",
+    },
+    divider: {
+        margin: "15px",
+    }
+})
+
+export { useStyles }
