+++ src/main/layout/index.js
@@ -1,5 +1,5 @@
-import Footer from "./Footer";
-import Header from "./Header";
+import Footer from "./footer/Footer";
+import Header from "./header/Header";
 import Navigation from "./Navigation";
 import Page from "./Page";
